refactor: combine feather and parquet datahandlers
